the first 98 Fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,20 +7,20 @@
  */
 int main(void)
 {
-	unsigned int a = 1;
-	unsigned int b = 2;
-	unsigned int c;
+	unsigned long a = 1;
+	unsigned long b = 2;
+	unsigned long c;
 	int count;
 
-	printf("%u, ", a);
-	printf("%u, ", b);
+	printf("%lu, ", a);
+	printf("%lu, ", b);
 	for (count = 0; count < 96; count++)
 	{
 		c = a + b;
 		if (count == 95)
-			printf("%u", c);
+			printf("%lu", c);
 		else
-			printf("%u, ", c);
+			printf("%lu, ", c);
 		a = b;
 		b = c;
 	}
